fixed dateTime parsing when reading file, created readFile function, created main_functions file so that all the basic functions could be in one place and easily modified
